requirements.txt
# run pip install -r requirements.txt to install all necessary dependencies
python-telegram-bot==22.1  # Telegram bot framework
python-dotenv>=1.0.0  # For loading environment variables from .env
python-telegram-bot[job-queue]
APScheduler==3.10.4